+++ setup/Cargo.toml
@@ -7,9 +7,3 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-drivers = { path = "../drivers"}
-async-stream = "0.3.0"
-futures-core = "0.3.12"
-futures-util = "0.3.12"
-tokio = { version = "1.0", features = ["rt-multi-thread", "macros", "sync", "time"] }
-async-std = "1.9.0"
